carservice/requirements.txt
Babel==2.7.0
cairocffi==1.0.2
CairoSVG==2.4.0
cffi==1.12.3
cssselect2==0.2.1
defusedxml==0.6.0
Django==2.2.1
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-fontawesome-5==1.0.16
django-phonenumber-field==3.0.1
djangorestframework==3.9.4
html5lib==1.0.1
phonenumbers==8.10.12
Pillow==6.0.0
psycopg2==2.8.2
pycparser==2.19
Pyphen==0.9.5
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
tinycss2==1.0.2
WeasyPrint==47
webencodings==0.5.1
